+++ src/index.ts
@@ -10,7 +10,9 @@ export * from './components/card';
 export * from './components/modal';
 export * from './components/toggle';
 export * from './components/option-menu';
+export * from './components/list';
 export * from './components/bottom-nav-bar';
+
 
 // New components
 export * from './components/accordion';